--- src/main/kotlin/com/example/estdelivery/adapters/in/web/dto/request/MemberCommand.kt
@@ -1,9 +1,0 @@
-package com.example.estdelivery.adapters.`in`.web.dto.request
-
-import com.example.estdelivery.domain.Member
-
-data class MemberCommand(
-    val name: String = ""
-)
-
-fun MemberCommand.toCreateDomain() = Member(name = name)
