+++ src/main/kotlin/com/example/estdelivery/adapters/in/web/member/dto/request/MemberCommand.kt
@@ -1,0 +1,9 @@
+package com.example.estdelivery.adapters.`in`.web.member.dto.request
+
+import com.example.estdelivery.domain.member.Member
+
+data class MemberCommand(
+    val name: String = ""
+)
+
+fun MemberCommand.toCreateDomain() = Member(name = name)
