+++ src/main/kotlin/com/example/estdelivery/application/port/in/member/CreateMemberUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.estdelivery.application.port.`in`.member
 
-import com.example.estdelivery.adapters.`in`.web.dto.request.MemberCommand
+import com.example.estdelivery.adapters.`in`.web.member.dto.request.MemberCommand
 import com.example.estdelivery.application.dto.member.MemberResponse
 
 interface CreateMemberUseCase {
