+++ src/main/kotlin/com/example/estdelivery/application/service/MemberService.kt
@@ -1,7 +1,7 @@
 package com.example.estdelivery.application.service
 
-import com.example.estdelivery.adapters.`in`.web.dto.request.MemberCommand
-import com.example.estdelivery.adapters.`in`.web.dto.request.toCreateDomain
+import com.example.estdelivery.adapters.`in`.web.member.dto.request.MemberCommand
+import com.example.estdelivery.adapters.`in`.web.member.dto.request.toCreateDomain
 import com.example.estdelivery.application.dto.member.MemberResponse
 import com.example.estdelivery.application.port.`in`.member.CreateMemberUseCase
 import com.example.estdelivery.application.port.out.member.MemberPersistencePort
